requirements.txt
streamlit==1.39.0
pyserial==3.5
matplotlib
pandas
numpy
pandas
plotly
